# Make `neighbors` for graphs faster by returning an iterator