--- a/src/graphs/BasicGraphs.jl
+++ b/src/graphs/BasicGraphs.jl
@@ -198,11 +198,11 @@ multiplicity*. To get the unique neighbors, call `unique(neighbors(g))`.
 
 """ In-neighbors of vertex in a graph.
 """
-@inline inneighbors(g::AbstractGraph, v::Int) = @inbounds subpart(g, incident(g, v, :tgt), :src)
+inneighbors(g::AbstractGraph, v::Int) = view(g, incident(g, v, :tgt), :src)
 
 """ Out-neighbors of vertex in a graph.
 """
-@inline outneighbors(g::AbstractGraph, v::Int) = @inbounds subpart(g, incident(g, v, :src), :tgt)
+outneighbors(g::AbstractGraph, v::Int) = view(g, incident(g, v, :src), :tgt)
 
 """ Union of in-neighbors and out-neighbors in a graph.
 """
@@ -285,11 +285,10 @@ rem_edge!(g::AbstractSymmetricGraph, e::Int) = rem_edges!(g, e:e)
 rem_edges!(g::AbstractSymmetricGraph, es) =
   rem_parts!(g, :E, unique!(sort!([es; inv(g, es)])))
 
-neighbors(g::AbstractSymmetricGraph, v::Int) =
-  subpart(g, incident(g, v, :src), :tgt)
-inneighbors(g::AbstractSymmetricGraph, v::Int) = neighbors(g, v)
-outneighbors(g::AbstractSymmetricGraph, v::Int) = neighbors(g, v)
-all_neighbors(g::AbstractSymmetricGraph, v::Int) = neighbors(g, v)
+neighbors(g::AbstractSymmetricGraph, v::Int) = view(g, incident(g, v, :src), :tgt)
+@inline inneighbors(g::AbstractSymmetricGraph, v::Int) = neighbors(g, v)
+@inline outneighbors(g::AbstractSymmetricGraph, v::Int) = neighbors(g, v)
+@inline all_neighbors(g::AbstractSymmetricGraph, v::Int) = neighbors(g, v)
 
 # Reflexive graphs
 ##################